move waypoint from secure_storage_config as a method of WaypointConfig
