+++ consensus/safety-rules/src/safety_rules_manager.rs
@@ -12,7 +12,7 @@ use crate::{
     SafetyRules, TSafetyRules,
 };
 use consensus_types::common::Author;
-use libra_config::config::{waypoint, NodeConfig, SafetyRulesService};
+use libra_config::config::{NodeConfig, SafetyRulesService};
 use libra_secure_storage::Storage;
 use std::{
     convert::TryInto,
@@ -37,7 +37,7 @@ pub fn extract_service_inputs(config: &mut NodeConfig) -> (Author, PersistentSaf
             .expect("Missing consensus keypair in test config")
             .take_private()
             .expect("Failed to take Consensus private key, key absent or already read");
-        let waypoint = waypoint(&config.base.waypoint);
+        let waypoint = config.base.waypoint.waypoint();
 
         PersistentSafetyStorage::initialize(internal_storage, private_key, waypoint)
     } else {
